+++ pages/search.tsx
@@ -1,10 +1,7 @@
 import SearchResultsCards from "@/components/SearchResultsCards";
-import { useRouter } from "next/router";
-import Header from "@/components/Header";
 import Loading from "@/components/Loading";
 import { AskMeResultData } from "@/types";
 import useSWR, { Fetcher } from "swr";
-import { useState } from "react";
 import { GetServerSideProps, InferGetServerSidePropsType } from "next";
 
 const fetcher: Fetcher<AskMeResultData, string> = (q) =>
@@ -13,29 +10,12 @@ const fetcher: Fetcher<AskMeResultData, string> = (q) =>
 export default function Search({
   queryProp,
 }: InferGetServerSidePropsType<typeof getServerSideProps>) {
-  const router = useRouter();
-  const [query, setQuery] = useState(queryProp);
-
   const { data } = useSWR(`/api/results?q=${queryProp}`, fetcher, {
     revalidateOnFocus: false,
   });
 
-  const fetchNewQuery = (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    if (query) {
-      router.push({
-        pathname: "/search",
-        query: {
-          q: query,
-        },
-      });
-    }
-  };
-
   return (
     <div>
-      <Header query={query} setQuery={setQuery} fetchNewQuery={fetchNewQuery} />
-
       {data ? <SearchResultsCards results={data.documents} /> : <Loading />}
     </div>
   );
